refactored dashboard page
